split local files and navigraph chart pages
